requirements.txt
paho-mqtt
mysql-connector-python
python-dotenv
requests